task/incidents/pyproject.toml
[tool.poetry]
name = "task"
version = "0.1.0"
description = ""
authors = ["bokakeshelava <keshelavaboka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.2"
django-rest-framework = "^0.1.0"
postgres = "^4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
